+++ mtex.js
@@ -1,9 +1,0 @@
-function add(a, b) {
-  return a + b;
-}
-
-let a = 1;
-let b = 2;
-
-console.log(add(a, b));
-